Update check_for_terms.py
Revised error-catching process to catch pdfminer.six errors and warnings in addition to empty and missing files. A new function (can_extract_text) returns a Boolean for the presence of any of the error/warning/problem conditions. The algorithm runs the function on each file and appends #NA values for Excel for files that return False from the function.